Remove unused handleNewMessage code
As long as we rely on socket.io to emit all the time back to us, including our own messages (eg. dont use broadcast), we shouldnt need this code. This was made when we thought socket.io wouldnt tell us to refresh the page, and so we'd have to update state for that single message